Features/giang/authen author (#20)
* fix auth controller conflict

* Features/giang/notification email (#12) conflict

* add func of notification when user have comment

* Add notification of new post to QA

* Change QsA email

* refactor: 💡 refactor some api endpoint and change some code (#13) conflict

* fix auth controller

* add permission conflict

* change something to fix pull

* add permission to functianal

* fix auth controller export json

* Add public file

* Push file vercel to setup

---------
