+++ README.md
@@ -1,10 +1,10 @@
 # Comp1640-BE
 
-<!-- PORT=
-MONGO_URL=
-SALT=
-SECRET_KEY=
-NODEMAILER_EMAIL=
-NODEMAILER_PASSWORD=
-APP_URL=
-QsA_EMAIL=-->
+<!-- PORT=3000
+MONGO_URL=mongodb+srv://thiengoten:Kn6w1GYUsmCeEsR3@comp1640.cujxxmk.mongodb.net/?retryWrites=true&w=majority
+SALT=10
+SECRET_KEY=concacnhan
+NODEMAILER_EMAIL=legiang1199@gmail.com
+NODEMAILER_PASSWORD=anilrlqjptoiimqq
+APP_URL='http://localhost:3000/'
+QsA_EMAIL=legiang1199@gmail.com -->
