+++ routes/api/bookmarks.js
@@ -1,11 +1,12 @@
 const express = require('express');
 const router = express.Router();
 const { createBookmark, deleteBookmark } = require('../../controllers/BookmarksController');
+const { userRole } = require('../../permission/author');
 
 //METHOD POST
-router.post('/', createBookmark);
+router.post('/', userRole, createBookmark);
 
 //METHOD DELETE
-router.delete('/:id', deleteBookmark);
+router.delete('/:id', userRole, deleteBookmark);
 
 module.exports = router;
