+++ routes/api/user.js
@@ -2,46 +2,28 @@ var express = require('express');
 var router = express.Router();
 
 const {
-  login,
-  register,
   getUserById,
   patchEditUser,
   deleteUser,
   getAllUser,
   uploadAvatar,
-  logout,
-  verifyEmail,
-  forgotPassword,
-  changePassword,
 } = require('../../controllers/UserController');
 
 const { avatarUploader } = require('../../middlewares/Uploader');
+const { adminRole } = require('../../permission/author');
 
 //METHOD GET
-router.get('/', getAllUser);
+router.get('/', adminRole, getAllUser);
 
 router.get('/:userId', getUserById);
-
-//METHOD POST
-router.post('/register', register);
-
-router.get('/verify-email/:userId', verifyEmail);
-
-router.post('/login', login);
-
-router.post('/logout', logout);
-
-router.post('/forgot-password', forgotPassword);
-
-router.post('/reset-password/:userId', changePassword);
 
 //upload user images
 router.post('/avatar/:userId', avatarUploader, uploadAvatar);
 
 //METHOD PATCH
-router.patch('/:userId', patchEditUser);
+router.patch('/:userId', adminRole, patchEditUser);
 
 //METHOD DELETE
-router.delete('/:userId', deleteUser);
+router.delete('/:userId', adminRole, deleteUser);
 
 module.exports = router;
